Keep old serializer separate
