+++ src/Speckle.Sdk.Dependencies/ArrayBufferWriterPooledObjectPolicy.cs
@@ -1,5 +1,4 @@
 ﻿using System.Buffers;
-using System.Text;
 using Microsoft.Extensions.ObjectPool;
 
 namespace Speckle.Sdk.Dependencies;
@@ -13,7 +12,7 @@ public sealed class ArrayBufferWriterPooledObjectPolicy<T> : PooledObjectPolicy<
   public int InitialCapacity { get; set; } = 100;
 
   /// <summary>
-  /// Gets or sets the maximum value for <see cref="StringBuilder.Capacity"/> that is allowed to be
+  /// Gets or sets the maximum value for <see cref="ArrayBufferWriter{T}.Capacity"/> that is allowed to be
   /// retained, when <see cref="Return(ArrayBufferWriter{T})"/> is invoked.
   /// </summary>
   /// <value>Defaults to <c>4096</c>.</value>
