+++ src/Speckle.Sdk/Serialisation/IdGenerator.cs
@@ -1,7 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using System.Diagnostics.Contracts;
-using System.Security.Cryptography;
-using System.Text;
+﻿using System.Diagnostics.Contracts;
 using Speckle.Sdk.Common;
 using Speckle.Sdk.Models;
 
@@ -19,43 +16,5 @@ public static class IdGenerator
     string hash = Sha256.GetString(serialized.Value, length: HashUtility.HASH_LENGTH);
 #endif
     return new Id(hash);
-  }
-
-#if NET6_0_OR_GREATER
-  [Pure]
-  public static string ComputeId(ReadOnlySpan<byte> input)
-  {
-    Span<byte> hash = stackalloc byte[SHA256.HashSizeInBytes];
-    SHA256.HashData(input, hash);
-
-    Span<char> output = stackalloc char[HashUtility.HASH_LENGTH];
-
-    for (int i = 0, j = 0; j < HashUtility.HASH_LENGTH; i += sizeof(byte), j += sizeof(char))
-    {
-      hash[i].TryFormat(output[j..], out _, "x2");
-    }
-
-    return new string(output);
-  }
-#endif
-
-  [Pure]
-  [SuppressMessage(
-    "Performance",
-    "CA1850:Prefer static \'HashData\' method over \'ComputeHash\'",
-    Justification = "We expose another overload with higher performance spans"
-  )]
-  public static string ComputeId(byte[] input, int offset, int count)
-  {
-    using var sha256 = SHA256.Create();
-    byte[] hash = sha256.ComputeHash(input, offset, count);
-
-    StringBuilder sb = new(64);
-    foreach (byte b in hash)
-    {
-      sb.Append(b.ToString("x2"));
-    }
-
-    return sb.ToString(0, HashUtility.HASH_LENGTH);
   }
 }
