+++ src/Speckle.Sdk/Serialisation/SerializationResult.cs
@@ -1,7 +1,4 @@
-﻿using System.Buffers;
-using Speckle.Sdk.Dependencies;
-
-namespace Speckle.Sdk.Serialisation;
+﻿namespace Speckle.Sdk.Serialisation;
 
 public readonly record struct SerializationResult(Json Json, Id? Id);
 
@@ -15,25 +12,6 @@ public readonly record struct Json
   public override string ToString() => Value;
 
   public string Value { get; }
-}
-
-public sealed class EfficientJson : IDisposable
-{
-  internal IBufferWriter<byte> Buffer => _value;
-
-  public ReadOnlySpan<byte> WrittenSpan => _value.WrittenSpan;
-
-  public ReadOnlyMemory<byte> WrittenMemory => _value.WrittenMemory;
-
-#if !NET5_0_OR_GREATER
-  public byte[] GetInternalBuffer() => _value.InternalBuffer;
-#endif
-
-  public int WrittenCount => _value.WrittenCount;
-
-  private readonly ArrayBufferWriter<byte> _value = Pools.ArrayBufferWriter.Get();
-
-  public void Dispose() => Pools.ArrayBufferWriter.Return(_value);
 }
 
 public readonly record struct Id
